added tests to validate test data against schema.  unit tests now passing but coverage low
