+++ app/uk/gov/hmrc/common/rest/resources/core/Declaration.scala
@@ -17,8 +17,7 @@
 package uk.gov.hmrc.common.rest.resources.core
 
 import org.joda.time.DateTime
-import play.api.libs.json.{Json, Reads}
-
+import play.api.libs.json.{JsString, Json, Reads, Writes}
 
 case class Declaration(correspondenceAddress: Address,
                        confirmation: Boolean,
@@ -30,5 +29,6 @@ case class Declaration(correspondenceAddress: Address,
 
 object Declaration {
   implicit val dateReads: Reads[DateTime] = Reads.of[String] map (new DateTime(_))
+  implicit val dateWrites: Writes[DateTime] = Writes { (dt: DateTime) => JsString(dt.toString("yyyy-MM-dd")) }
   implicit val formats = Json.format[Declaration]
 }
