+++ public/api/trusts/conf/1.0/examples/trustees.json
@@ -8,7 +8,7 @@
       "dateOfBirth": "1989-12-31",
       "nino": "WA123456A",
       "passport": {
-        "identifier": "b12344b23b45",
+        "referenceNumber": "b12344b23b45",
         "expiryDate": "2018-12-25",
         "countryOfIssue": "UK"
       },
@@ -31,7 +31,7 @@
       "dateOfBirth": "1989-12-31",
       "nino": "WA123456A",
       "passport": {
-        "identifier": "b12344b23b45",
+        "referenceNumber": "b12344b23b45",
         "expiryDate": "2018-12-25",
         "countryOfIssue": "UK"
       },
