+++ test/resources/ValidPassport.json
@@ -1,5 +1,5 @@
 {
-  "identifier": "IDENTIFIER",
+  "referenceNumber": "IDENTIFIER",
   "expiryDate": "2020-01-01",
   "countryOfIssue": "ES"
 }
