+++ test/uk/gov/hmrc/common/DESconnector/TrustExistenceSpec.scala
@@ -21,7 +21,7 @@ import org.mockito.Mockito.when
 import org.scalatestplus.play.{OneAppPerSuite, PlaySpec}
 import uk.gov.hmrc.play.http.{HttpResponse, Upstream4xxResponse}
 import uk.gov.hmrc.trustapi.rest.resources.core.TrustExistence
-import uk.gov.hmrc.trustregistration.ScalaDataExamples
+import uk.gov.hmrc.utils.ScalaDataExamples
 
 import scala.concurrent.duration.Duration
 import scala.concurrent.{Await, Future}
