+++ test/uk/gov/hmrc/estateapi/rest/resources/EstateSpec.scala
@@ -17,8 +17,7 @@
 package uk.gov.hmrc.trustregistration.models
 
 import org.scalatestplus.play.PlaySpec
-import uk.gov.hmrc.trustregistration.ScalaDataExamples
-
+import uk.gov.hmrc.utils.ScalaDataExamples
 
 
 class EstateSpec extends PlaySpec with ScalaDataExamples {
