+++ test/uk/gov/hmrc/trustapi/rest/resources/LeadTrusteeSpec.scala
@@ -18,7 +18,7 @@ package uk.gov.hmrc.trustregistration.models
 
 import org.scalatestplus.play.PlaySpec
 import uk.gov.hmrc.trustapi.rest.resources.core.LeadTrustee
-import uk.gov.hmrc.trustregistration.ScalaDataExamples
+import uk.gov.hmrc.utils.ScalaDataExamples
 
 class LeadTrusteeSpec extends PlaySpec with ScalaDataExamples {
