+++ test/uk/gov/hmrc/trustapi/rest/resources/ProtectorsSpec.scala
@@ -19,7 +19,7 @@ package uk.gov.hmrc.trustregistration.models
 import org.scalatestplus.play.PlaySpec
 import play.api.libs.json.Json
 import uk.gov.hmrc.trustapi.rest.resources.core.Protectors
-import uk.gov.hmrc.trustregistration.{JsonExamples, ScalaDataExamples}
+import uk.gov.hmrc.utils.{JsonExamples, ScalaDataExamples}
 
 class ProtectorsSpec extends PlaySpec with JsonExamples with ScalaDataExamples {
