+++ test/uk/gov/hmrc/trustapi/rest/resources/TrustTypeSpec.scala
@@ -18,7 +18,7 @@ package uk.gov.hmrc.trustregistration.models
 
 import org.scalatestplus.play.PlaySpec
 import uk.gov.hmrc.trustapi.rest.resources.core.trusttypes.TrustType
-import uk.gov.hmrc.trustregistration.ScalaDataExamples
+import uk.gov.hmrc.utils.ScalaDataExamples
 
 class TrustTypeSpec extends PlaySpec with ScalaDataExamples {
   "TrustType" must {
