+++ test/uk/gov/hmrc/trustapi/rest/resources/WillIntestacyTrustSpec.scala
@@ -20,7 +20,7 @@ import org.scalatestplus.play.PlaySpec
 import uk.gov.hmrc.trustapi.rest.resources.core.assets.Assets
 import uk.gov.hmrc.trustapi.rest.resources.core.beneficiaries.Beneficiaries
 import uk.gov.hmrc.trustapi.rest.resources.core.trusttypes.WillIntestacyTrust
-import uk.gov.hmrc.trustregistration.ScalaDataExamples
+import uk.gov.hmrc.utils.ScalaDataExamples
 
 class WillIntestacyTrustSpec extends PlaySpec with ScalaDataExamples {
